Emazon/src/main/java/com/bootcamp_2024_2/emazon/domain/model/CustomPageable.java
package com.bootcamp_2024_2.emazon.domain.model;

public class CustomPageable {
    private int pageNumber;
    private int pageSize;
    private String sortBy;
    private String sortOrder;

    public CustomPageable(int pageNumber, int pageSize, String sortBy, String sortOrder) {
        this.pageNumber = pageNumber;
        this.pageSize = pageSize;
        this.sortBy = sortBy;
        this.sortOrder = sortOrder;
    }

    public CustomPageable() {
    }

    public int getPageNumber() {
        return pageNumber;
    }

    public void setPageNumber(int pageNumber) {
        this.pageNumber = pageNumber;
    }

    public int getPageSize() {
        return pageSize;
    }

    public void setPageSize(int pageSize) {
        this.pageSize = pageSize;
    }

    public String getSortBy() {
        return sortBy;
    }

    public void setSortBy(String sortBy) {
        this.sortBy = sortBy;
    }

    public String getSortOrder() {
        return sortOrder;
    }

    public void setSortOrder(String sortOrder) {
        this.sortOrder = sortOrder;
    }
}
